custom loss function

+++ model.py
@@ -104,7 +104,7 @@ def build_model(config):
     def custom_loss(y_true, y_pred):
         main_loss = losses.sparse_categorical_crossentropy(y_true, y_pred)
         pred_indices = K.argmax(y_pred, axis=-1)
-        distance_penalty = K.constant(1.0) / K.abs(pred_indices - K.constant(config.n_classes / 2))
+        distance_penalty = K.constant(1.0, dtype='float32') / K.abs(pred_indices - K.constant(config.n_classes / 2, dtype='float32'))
         return main_loss + config.distance_weight * distance_penalty
 
     opt = Adam(lr=config.lr)
